+++ app/flask_app/models/embedding/prepare_model.sh
